D3ASIM-3307: Refactor simulation results classes to use a common abstract base class to adhere to the same interface. Moved the merge results methods to their respective classes.
